cleaned up os thingy

+++ nsl.h
@@ -144,6 +144,7 @@ typedef enum {
     NSL_ERROR_IS_DIRECTORY,
     NSL_ERROR_FILE_BUSY,
     NSL_ERROR_PARSE,
+    NSL_ERROR_PATH_TOO_LONG,
 
     NSL_ERROR = -1,
 } nsl_Error;
@@ -1213,9 +1214,13 @@ NSL_API Function nsl_dll_symbol(nsl_Dll *handle, nsl_Str symbol) {
 #include <errno.h>
 #include <unistd.h>
 
+#define NSL_OS_PATH_MAX FILENAME_MAX
+
 NSL_API nsl_Error nsl_os_mkdir_conf(nsl_Path path, nsl_OsDirConfig config) {
+    if (path.len >= NSL_OS_PATH_MAX - 1) return NSL_ERROR_PATH_TOO_LONG;
+
     if (config.parents) {
-        if (nsl_path_is_root(path)) return NSL_NO_ERROR;
+        if (nsl_path_is_root(path))               return NSL_NO_ERROR;
         if (path.len == 1 && path.data[0] == '.') return NSL_NO_ERROR;;
         nsl_OsDirConfig c = config;
         c.exists_ok = true;
@@ -1224,8 +1229,8 @@ NSL_API nsl_Error nsl_os_mkdir_conf(nsl_Path path, nsl_OsDirConfig config) {
     }
 
     errno = 0;
-    char filepath[FILENAME_MAX] = {0};
-    memcpy(filepath, path.data, nsl_usize_min(path.len, FILENAME_MAX - 1));
+    char filepath[NSL_OS_PATH_MAX] = {0};
+    memcpy(filepath, path.data, nsl_usize_min(path.len, NSL_OS_PATH_MAX - 1));
     if (mkdir(filepath, config.mode ? config.mode : 0755) != 0) {
         if (config.exists_ok && errno == EEXIST) {
             struct stat info;
@@ -1240,9 +1245,11 @@ NSL_API nsl_Error nsl_os_mkdir_conf(nsl_Path path, nsl_OsDirConfig config) {
 }
 
 NSL_API nsl_Error nsl_os_chdir(nsl_Path path) {
+    if (path.len >= NSL_OS_PATH_MAX - 1) return NSL_ERROR_PATH_TOO_LONG;
+
     errno = 0;
-    char filepath[FILENAME_MAX] = {0};
-    memcpy(filepath, path.data, nsl_usize_min(path.len, FILENAME_MAX - 1));
+    char filepath[NSL_OS_PATH_MAX] = {0};
+    memcpy(filepath, path.data, nsl_usize_min(path.len, NSL_OS_PATH_MAX - 1));
     if (chdir(filepath) != 0) {
         if (errno == EACCES)  return NSL_ERROR_ACCESS_DENIED;
         if (errno == ENOENT)  return NSL_ERROR_FILE_NOT_FOUND;
@@ -1270,14 +1277,18 @@ NSL_API nsl_Str nsl_os_getenv(const char *env, nsl_Arena* arena) {
 }
 
 NSL_API bool nsl_os_exists(nsl_Path path) {
-    char filepath[FILENAME_MAX] = {0};
-    memcpy(filepath, path.data, nsl_usize_min(path.len, FILENAME_MAX - 1));
+    if (path.len >= NSL_OS_PATH_MAX - 1) return false;
+
+    char filepath[NSL_OS_PATH_MAX] = {0};
+    memcpy(filepath, path.data, nsl_usize_min(path.len, NSL_OS_PATH_MAX - 1));
     return access(filepath, 0) == 0;
 }
 
 NSL_API bool nsl_os_is_dir(nsl_Path path) {
-    char filepath[FILENAME_MAX] = {0};
-    memcpy(filepath, path.data, nsl_usize_min(path.len, FILENAME_MAX - 1));
+    if (path.len >= NSL_OS_PATH_MAX - 1) return NSL_ERROR_PATH_TOO_LONG;
+
+    char filepath[NSL_OS_PATH_MAX] = {0};
+    memcpy(filepath, path.data, nsl_usize_min(path.len, NSL_OS_PATH_MAX - 1));
 
     struct stat info;
     if (stat(filepath, &info) == -1) {
@@ -1288,8 +1299,10 @@ NSL_API bool nsl_os_is_dir(nsl_Path path) {
 }
 
 NSL_API nsl_Error nsl_os_remove(nsl_Path path) {
-    char filepath[FILENAME_MAX] = {0};
-    memcpy(filepath, path.data, nsl_usize_min(path.len, FILENAME_MAX - 1));
+    if (path.len >= NSL_OS_PATH_MAX - 1) return NSL_ERROR_PATH_TOO_LONG;
+
+    char filepath[NSL_OS_PATH_MAX] = {0};
+    memcpy(filepath, path.data, nsl_usize_min(path.len, NSL_OS_PATH_MAX - 1));
 
     errno = 0;
     if (unlink(filepath) != 0) {
@@ -1304,10 +1317,10 @@ NSL_API nsl_Error nsl_os_remove(nsl_Path path) {
 }
 
 NSL_API bool nsl_os_older_than(nsl_Path p1, nsl_Path p2) {
-    char filepath[FILENAME_MAX] = {0};
-    struct stat info[2];
+    if (p1.len >= NSL_OS_PATH_MAX - 1 || p1.len >= NSL_OS_PATH_MAX - 1) return false;
 
-    if (p1.len > FILENAME_MAX || p2.len > FILENAME_MAX) return false;
+    char filepath[NSL_OS_PATH_MAX] = {0};
+    struct stat info[2];
 
     memcpy(filepath, p1.data, p1.len);
     filepath[p1.len] = '\0';
@@ -1570,28 +1583,32 @@ NSL_API Function nsl_dll_symbol(nsl_Dll *dll, nsl_Str symbol) {
 #include <string.h>
 #include <errno.h>
 
+#define NSL_OS_PATH_MAX PATH_MAX
+
 NSL_API nsl_Error nsl_os_mkdir_(nsl_Path path, nsl_OsDirConfig config) {
-    nsl_Error result = NSL_NO_ERROR;
+    if (path.len >= NSL_OS_PATH_MAX) return NSL_ERROR_PATH_TOO_LONG;
 
     nsl_Arena arena = {0};
 
     if (config.parents) {
-        if (nsl_path_is_root(path)) NSL_DEFER(NSL_NO_ERROR);
-        if (path.len == 1 && path.data[0] == '.') NSL_DEFER(NSL_NO_ERROR);
+        if (nsl_path_is_root(path))               return NSL_NO_ERROR;
+        if (path.len == 1 && path.data[0] == '.') return NSL_NO_ERROR;
         nsl_OsDirConfig c = config;
         c.exists_ok = true;
         nsl_Error recursive_error = nsl_os_mkdir_(nsl_path_parent(path), c);
-        if (recursive_error) NSL_DEFER(recursive_error);
+        if (recursive_error) return recursive_error;
     }
 
-    nsl_Str filepath = nsl_str_copy(path, &arena);
+    char filepath[NSL_OS_PATH_MAX] = {0};
+    memcpy(filepath, path.data, path.len);
+    filepath[path.len] = '\0';
 
-    if (!CreateDirectoryA(filepath.data, NULL)) {
+    if (!CreateDirectoryA(filepath, NULL)) {
         DWORD ec = GetLastError();
         if (config.exists_ok && ec == ERROR_ALREADY_EXISTS) {
-            DWORD attrs = GetFileAttributes(filepath.data);
+            DWORD attrs = GetFileAttributes(filepath);
             if (attrs != INVALID_FILE_ATTRIBUTES && (attrs & FILE_ATTRIBUTE_DIRECTORY)) {
-                NSL_DEFER(NSL_NO_ERROR);
+                return NSL_NO_ERROR;
             }
         }
         if (ec == ERROR_ALREADY_EXISTS)     NSL_DEFER(NSL_ERROR_ALREADY_EXISTS);
@@ -1610,16 +1627,15 @@ NSL_API nsl_Error nsl_os_mkdir_(nsl_Path path, nsl_OsDirConfig config) {
         NSL_PANIC(msg);
     }
 
-defer:
-    nsl_arena_free(&arena);
-    return result;
+    return NSL_NO_ERROR;
 }
 
 NSL_API nsl_Error nsl_os_chdir(nsl_Path path) {
-    NSL_ASSERT(path.len >= MAX_PATH - 1 && "Filename is too long for windows to handle");
+    if (path.len >= NSL_OS_PATH_MAX - 1) return NSL_ERROR_PATH_TOO_LONG;
 
-    char pathname[MAX_PATH] = {0};
+    char pathname[NSL_OS_PATH_MAX] = {0};
     memcpy(pathname, path.data, path.len);
+    pathname[path.len] = '\0';
 
     if (!SetCurrentDirectoryA(pathname)) {
         DWORD ec = GetLastError();
@@ -1688,9 +1704,9 @@ NSL_API nsl_Str nsl_os_getenv(const char *env, nsl_Arena *arena) {
 }
 
 NSL_API bool nsl_os_exists(nsl_Path path) {
-    NSL_ASSERT(path.len < MAX_PATH && "pathname is too long for windows to handle");
+    if (path.len >= NSL_OS_PATH_MAX - 1) return false;
 
-    char filepath[MAX_PATH] = {0};
+    char filepath[NSL_OS_PATH_MAX] = {0};
     memcpy(filepath, path.data, path.len);
     DWORD dwAttrib = GetFileAttributesA(filepath);
 
@@ -1698,20 +1714,23 @@ NSL_API bool nsl_os_exists(nsl_Path path) {
 }
 
 NSL_API bool nsl_os_is_dir(nsl_Path path) {
-    NSL_ASSERT(path.len < MAX_PATH && "pathname is too long for windows to handle");
+    if (path.len >= NSL_OS_PATH_MAX - 1) return false;
 
-    char filepath[MAX_PATH] = {0};
+    char filepath[NSL_OS_PATH_MAX] = {0};
     memcpy(filepath, path.data, path.len);
+    filepath[path.len] = '\0';
     DWORD dwAttrib = GetFileAttributesA(filepath);
 
     return (dwAttrib != INVALID_FILE_ATTRIBUTES && (dwAttrib & FILE_ATTRIBUTE_DIRECTORY));
 }
 
 NSL_API nsl_Error nsl_os_remove(nsl_Path path) {
-    NSL_ASSERT(path.len < MAX_PATH && "pathname is too long for windows to handle");
+    if (path.len >= NSL_OS_PATH_MAX - 1) return NSL_ERROR_PATH_TOO_LONG;
 
-    char filepath[MAX_PATH] = {0};
+    char filepath[NSL_OS_PATH_MAX] = {0};
     memcpy(filepath, path.data, path.len);
+    filepath[path.len] = '\0';
+
     if (DeleteFileA(filepath) == 0) {
         DWORD ec = GetLastError();
         if (ec == ERROR_ACCESS_DENIED)     return NSL_ERROR_ACCESS_DENIED;
@@ -1734,11 +1753,11 @@ NSL_API nsl_Error nsl_os_remove(nsl_Path path) {
 }
 
 NSL_API bool nsl_os_older_than(nsl_Path p1, nsl_Path p2) {
-    char filepath[MAX_PATH] = {0};
+    char filepath[NSL_OS_PATH_MAX] = {0};
     FILETIME ft[2] = {0};
     HANDLE hFile[2] = {INVALID_HANDLE_VALUE, INVALID_HANDLE_VALUE};
 
-    if (p1.len > MAX_PATH || p2.len > MAX_PATH) return false;
+    if (p1.len > NSL_OS_PATH_MAX || p2.len > NSL_OS_PATH_MAX) return false;
 
     memcpy(filepath, p1.data, p1.len);
     filepath[p1.len] = '\0';
